add : auth.js & pagination 구현

+++ book-shop/controller/OrderController.js
@@ -1,47 +1,61 @@
-// const conn = require('../mariadb'); //db 모듈
 const mariadb = require('mysql2/promise'); //db 모듈
+const ensureAuthorization = require('../auth');
+const jwt = require('jsonwebtoken');
 const { StatusCodes } = require('http-status-codes'); //status code 모듈
 
 const order = async (req, res) => {
 
     const conn = await mariadb.createConnection({
-        host : '127.0.0.1',
-        user : 'root',
-        password : 'root',
-        database : 'Bookshop',
-        dataStrings : true
+        host: '127.0.0.1',
+        user: 'root',
+        password: 'root',
+        database: 'Bookshop',
+        dataStrings: true
     });
 
-    const { items, delivery, totalQuantity, totalPrice, userId, firstBookTitle } = req.body;
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = 'INSERT INTO delivery (address, receiver, contact) VALUES (?,?,?)';
-    let values = [delivery.address, delivery.receiver, delivery.contact];
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        })
+    } else {
 
-    let [results] = await conn.execute(sql, values); //실행한 결과값을 배열형태로 받아줌.. select에서는 [rows, fields]
-    
-    let delivery_id = results.insertId;
+        const { items, delivery, totalQuantity, totalPrice, firstBookTitle } = req.body;
 
-    sql = "INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) VALUES (?, ?, ?, ?, ?)";
-    values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-    [results] = await conn.execute(sql, values);
+        let sql = 'INSERT INTO delivery (address, receiver, contact) VALUES (?,?,?)';
+        let values = [delivery.address, delivery.receiver, delivery.contact];
 
-    let order_id = results.insertId;
+        let [results] = await conn.execute(sql, values); //실행한 결과값을 배열형태로 받아줌.. select에서는 [rows, fields]
 
-    // items를 가지고, 장바구니에서 book_id, quantity 조회
-    sql = 'SELECT book_id, quantity FROM cartItems WHERE id IN (?)';
-    let [orderItems, fields] = await conn.query(sql, [items]);
+        let delivery_id = results.insertId;
 
-    // orderedBook 테이블 삽입
-    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
-    //items.. 배열 : 요소들을 하나씩 꺼내서 (forEach문 돌려서)
-    values = [];
-    orderItems.forEach((item) => { values.push([order_id, item.book_id, item.quantity]) }
-    )
-    results = await conn.query(sql, [values]);
+        sql = "INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) VALUES (?, ?, ?, ?, ?)";
+        values = [firstBookTitle, totalQuantity, totalPrice, authorization.id, delivery_id];
+        [results] = await conn.execute(sql, values);
 
-    let result = await deleteCartItems(conn, items);
+        let order_id = results.insertId;
 
-    return res.status(StatusCodes.OK).json(result);
+        // items를 가지고, 장바구니에서 book_id, quantity 조회
+        sql = 'SELECT book_id, quantity FROM cartItems WHERE id IN (?)';
+        let [orderItems, fields] = await conn.query(sql, [items]);
+
+        // orderedBook 테이블 삽입
+        sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
+        //items.. 배열 : 요소들을 하나씩 꺼내서 (forEach문 돌려서)
+        values = [];
+        orderItems.forEach((item) => { values.push([order_id, item.book_id, item.quantity]) }
+        )
+        results = await conn.query(sql, [values]);
+
+        let result = await deleteCartItems(conn, items);
+
+        return res.status(StatusCodes.OK).json(result);
+    }
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -52,38 +66,62 @@ const deleteCartItems = async (conn, items) => {
 }
 
 const getOrders = async (req, res) => {
-    const conn = await mariadb.createConnection({
-        host : '127.0.0.1',
-        user : 'root',
-        password : 'root',
-        database : 'Bookshop',
-        dataStrings : true
-    });
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = `SELECT orders.id, created_at,  address, receiver, contact, book_title, total_quantity, total_price
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        })
+    } else {
+        const conn = await mariadb.createConnection({
+            host: '127.0.0.1',
+            user: 'root',
+            password: 'root',
+            database: 'Bookshop',
+            dataStrings: true
+        });
+
+        let sql = `SELECT orders.id, created_at,  address, receiver, contact, book_title, total_quantity, total_price
                 FROM orders LEFT JOIN delivery 
                 ON orders.delivery_id = delivery.id`;
-    let [rows, fields] = await conn.query(sql);
-    return res.status(StatusCodes.OK).json(rows);
+        let [rows, fields] = await conn.query(sql);
+        return res.status(StatusCodes.OK).json(rows);
+    }
 }
 
 const getOrderDetail = async (req, res) => {
-    const conn = await mariadb.createConnection({
-        host : '127.0.0.1',
-        user : 'root',
-        password : 'root',
-        database : 'Bookshop',
-        dataStrings : true
-    });
+    let authorization = ensureAuthorization(req, res);
 
-    const {id} = req.params;
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        })
+    } else {
+        const conn = await mariadb.createConnection({
+            host: '127.0.0.1',
+            user: 'root',
+            password: 'root',
+            database: 'Bookshop',
+            dataStrings: true
+        });
 
-    let sql = `SELECT book_id, title, author, price, quantity
+        const orderId = req.params.id;
+
+        let sql = `SELECT book_id, title, author, price, quantity
                 FROM orderedBook LEFT JOIN books 
                 ON orderedBook.book_id = books.id
                 WHERE order_id = ?`;
-    let [rows, fields] = await conn.query(sql, id);
-    return res.status(StatusCodes.OK).json(rows);
+        let [rows, fields] = await conn.query(sql, orderId);
+        return res.status(StatusCodes.OK).json(rows);
+    }
 }
 
 module.exports = {
